back/requirements.txt
fastapi==0.115.6
uvicorn==0.30.6
sqlalchemy==2.0.36
pydantic==2.10.3
python-jose==3.3.0
passlib[bcrypt]==1.7.4
email-validator==2.2.0
bcrypt==4.0.1
authlib==1.3.0
httpx==0.27.2
python-dotenv==1.0.1
itsdangerous==2.1.2
google-api-python-client==2.160.0
google-auth==2.36.0
google-auth-httplib2==0.2.0
